patternLab4/requirements.txt
requests
pandas
redis
kafka-python